Keep implementations of resource constructors in their files
