+++ docs/lib/framebuffer.d.ts
@@ -5,6 +5,7 @@ import { Texture } from "./texture";
 export declare type TextureColorInternalFormat = InternalFormat.R8 | InternalFormat.R8_SNORM | InternalFormat.R8UI | InternalFormat.R8I | InternalFormat.R16UI | InternalFormat.R16I | InternalFormat.R32UI | InternalFormat.R32I | InternalFormat.R16F | InternalFormat.R32F | InternalFormat.RG8 | InternalFormat.RG8_SNORM | InternalFormat.RG8UI | InternalFormat.RG8I | InternalFormat.RG16UI | InternalFormat.RG16I | InternalFormat.RG32UI | InternalFormat.RG32I | InternalFormat.RG16F | InternalFormat.RG32F | InternalFormat.RGB8 | InternalFormat.RGB8_SNORM | InternalFormat.RGB8UI | InternalFormat.RGB8I | InternalFormat.RGB16UI | InternalFormat.RGB16I | InternalFormat.RGB32UI | InternalFormat.RGB32I | InternalFormat.RGB16F | InternalFormat.RGB32F | InternalFormat.RGBA8 | InternalFormat.RGBA8_SNORM | InternalFormat.RGBA8UI | InternalFormat.RGBA8I | InternalFormat.RGBA16UI | InternalFormat.RGBA16I | InternalFormat.RGBA32UI | InternalFormat.RGBA32I | InternalFormat.RGBA16F | InternalFormat.RGBA32F;
 export declare type TextureDepthInternalFormat = InternalFormat.DEPTH_COMPONENT16 | InternalFormat.DEPTH_COMPONENT24 | InternalFormat.DEPTH_COMPONENT32F;
 export declare type TextureDepthStencilInternalFormat = InternalFormat.DEPTH24_STENCIL8 | InternalFormat.DEPTH32F_STENCIL8;
+export declare function _createFramebuffer(state: State, width: number, height: number, color: Texture<TextureColorInternalFormat> | Texture<TextureColorInternalFormat>[], depthStencil?: Texture<TextureDepthInternalFormat> | Texture<TextureDepthStencilInternalFormat>): Framebuffer;
 /**
  * Framebuffers store the list of attachments to write to during a draw
  * operation. They can be a draw target via `framebuffer.target()`
