+++ src/device.ts
@@ -1,34 +1,23 @@
-import * as assert from "./util/assert";
-import * as array from "./util/array";
 import {
-    BufferUsage,
     Primitive,
     DataType,
     Format,
     InternalFormat,
-    Filter,
-    Wrap,
-    sizeOf,
 } from "./types";
-import {
-    State,
-    DepthTestDescriptor,
-    StencilTestDescriptor,
-    BlendDescriptor,
-} from "./state";
+import { State } from "./state";
 import { Target } from "./target";
 import {
     Command,
     CommandCreateOptions,
-    DepthFunc,
-    StencilOp,
-    BlendEquation,
+    _createCommand,
 } from "./command";
 import {
     VertexBuffer,
     VertexBufferCreateOptions,
     VertexBufferType,
     VertexBufferTypeToTypedArray,
+    _createVertexBuffer,
+    _createVertexBufferWithTypedArray,
 } from "./vertex-buffer";
 import {
     ElementBuffer,
@@ -36,13 +25,15 @@ import {
     ElementArray,
     ElementBufferType,
     ElementBufferTypeToTypedArray,
+    _createElementBuffer,
+    _createElementBufferWithArray,
+    _createElementBufferWithTypedArray,
 } from "./element-buffer";
 import {
     Attributes,
     AttributesConfig,
     AttributesCreateOptions,
-    AttributeDescriptor,
-    AttributeType,
+    _createAttributes,
 } from "./attributes";
 import {
     Texture,
@@ -52,16 +43,17 @@ import {
     InternalFormatToTypedArray,
     InternalFormatToDataFormat,
     InternalFormatToDataType,
+    _createTexture,
+    _createTextureWithTypedArray,
 } from "./texture";
 import {
     Framebuffer,
     TextureColorInternalFormat,
     TextureDepthInternalFormat,
     TextureDepthStencilInternalFormat,
+    _createFramebuffer,
 } from "./framebuffer";
 
-
-const INT_PATTERN = /^0|[1-9]\d*$/;
 
 /**
  * Available extensions.
@@ -314,31 +306,9 @@ export class Device {
     createCommand<P = void>(
         vert: string,
         frag: string,
-        {
-            textures = {},
-            uniforms = {},
-            depth,
-            stencil,
-            blend,
-        }: CommandCreateOptions<P> = {},
+        options?: CommandCreateOptions<P>,
     ): Command<P> {
-        assert.nonNull(vert, fmtParamNonNull("vert"));
-        assert.nonNull(frag, fmtParamNonNull("frag"));
-
-        const depthDescr = parseDepth(depth);
-        const stencilDescr = parseStencil(stencil);
-        const blendDescr = parseBlend(blend);
-
-        return new Command(
-            this.state,
-            vert,
-            frag,
-            textures,
-            uniforms,
-            depthDescr,
-            stencilDescr,
-            blendDescr,
-        );
+        return _createCommand(this.state, vert, frag, options);
     }
 
     /**
@@ -347,15 +317,9 @@ export class Device {
     createVertexBuffer<T extends VertexBufferType>(
         type: T,
         size: number,
-        { usage = BufferUsage.DYNAMIC_DRAW }: VertexBufferCreateOptions = {},
+        options?: VertexBufferCreateOptions,
     ): VertexBuffer<T> {
-        return new VertexBuffer(
-            this._gl,
-            type,
-            size,
-            size * sizeOf(type),
-            usage,
-        );
+        return _createVertexBuffer(this._gl, type, size, options);
     }
 
     /**
@@ -365,35 +329,26 @@ export class Device {
     createVertexBufferWithTypedArray<T extends VertexBufferType>(
         type: T,
         data: VertexBufferTypeToTypedArray[T] | number[],
-        { usage = BufferUsage.STATIC_DRAW }: VertexBufferCreateOptions = {},
+        options?: VertexBufferCreateOptions,
     ): VertexBuffer<T> {
-        return new VertexBuffer(
+        return _createVertexBufferWithTypedArray(
             this._gl,
             type,
-            data.length,
-            data.length * sizeOf(type),
-            usage,
-        ).store(data);
+            data,
+            options,
+        );
     }
 
     /**
      * Create a new element buffer with given type, primitive, and size.
      */
     createElementBuffer<T extends ElementBufferType>(
-        dev: Device,
         type: T,
         primitive: Primitive,
         size: number,
-        { usage = BufferUsage.DYNAMIC_DRAW }: ElementBufferCreateOptions = {},
+        options?: ElementBufferCreateOptions,
     ): ElementBuffer<T> {
-        return new ElementBuffer(
-            dev._gl,
-            type,
-            primitive,
-            size,
-            size * sizeOf(type),
-            usage,
-        );
+        return _createElementBuffer(this._gl, type, primitive, size, options);
     }
 
     /**
@@ -408,27 +363,7 @@ export class Device {
         data: ElementArray,
         options?: ElementBufferCreateOptions,
     ): ElementBuffer<DataType.UNSIGNED_INT> {
-        if (array.is2(data)) {
-            const shape = array.shape2(data);
-            assert.rangeInclusive(shape[1], 2, 3, (p) => {
-                return `Elements must be 2-tuples or 3-tuples, got ${p}-tuple`;
-            });
-            const ravel = array.ravel2(data, shape);
-            const primitive = shape[1] === 3
-                ? Primitive.TRIANGLES
-                : Primitive.LINES;
-            return this.createElementBufferWithTypedArray(
-                DataType.UNSIGNED_INT,
-                primitive,
-                ravel,
-            );
-        }
-        return this.createElementBufferWithTypedArray(
-            DataType.UNSIGNED_INT,
-            Primitive.POINTS,
-            data,
-            options,
-        );
+        return _createElementBufferWithArray(this._gl, data, options);
     }
 
     /**
@@ -439,16 +374,15 @@ export class Device {
         type: T,
         primitive: Primitive,
         data: ElementBufferTypeToTypedArray[T] | number[],
-        { usage = BufferUsage.STATIC_DRAW }: ElementBufferCreateOptions = {},
+        options?: ElementBufferCreateOptions,
     ): ElementBuffer<T> {
-        return new ElementBuffer(
+        return _createElementBufferWithTypedArray(
             this._gl,
             type,
             primitive,
-            data.length,
-            data.length * sizeOf(type),
-            usage,
-        ).store(data);
+            data,
+            options,
+        );
     }
 
     /**
@@ -469,106 +403,9 @@ export class Device {
     createAttributes(
         elements: Primitive | ElementArray | ElementBuffer<ElementBufferType>,
         attributes: AttributesConfig,
-        { countLimit }: AttributesCreateOptions = {},
+        options?: AttributesCreateOptions,
     ): Attributes {
-        if (typeof countLimit === "number") {
-            assert.gt(countLimit, 0, (p) => {
-                return `Count limit must be greater than 0, got: ${p}`;
-            });
-        }
-
-        const attrs = Object.entries(attributes)
-            .map(([locationStr, definition]) => {
-                if (!INT_PATTERN.test(locationStr)) {
-                    throw new Error("Location not a number. Use Command#locate");
-                }
-                const location = parseInt(locationStr, 10);
-                if (Array.isArray(definition)) {
-                    if (array.is2(definition)) {
-                        const s = array.shape2(definition);
-                        const r = array.ravel2(definition, s);
-                        return new AttributeDescriptor(
-                            location,
-                            AttributeType.POINTER,
-                            this.createVertexBufferWithTypedArray(
-                                DataType.FLOAT,
-                                r,
-                            ),
-                            s[0],
-                            s[1],
-                            false,
-                            0,
-                        );
-                    }
-                    return new AttributeDescriptor(
-                        location,
-                        AttributeType.POINTER,
-                        this.createVertexBufferWithTypedArray(
-                            DataType.FLOAT,
-                            definition,
-                        ),
-                        definition.length,
-                        1,
-                        false,
-                        0,
-                    );
-                }
-
-                return new AttributeDescriptor(
-                    location,
-                    definition.type,
-                    Array.isArray(definition.buffer)
-                        ? this.createVertexBufferWithTypedArray(
-                            DataType.FLOAT,
-                            definition.buffer,
-                        )
-                        : definition.buffer,
-                    definition.count,
-                    definition.size,
-                    definition.type === AttributeType.POINTER
-                        ? (definition.normalized || false)
-                        : false,
-                    definition.divisor || 0,
-                );
-            });
-
-        let primitive: Primitive;
-        let elementBuffer: ElementBuffer<ElementBufferType> | undefined;
-        if (typeof elements === "number") {
-            primitive = elements;
-        } else {
-            elementBuffer = elements instanceof ElementBuffer
-                ? elements
-                : this.createElementBufferWithArray(elements);
-            primitive = elementBuffer.primitive;
-        }
-
-        const inferredCount = elementBuffer
-            ? elementBuffer.length
-            : attrs.length
-                ? attrs
-                    .map((attr) => attr.count)
-                    .reduce((min, curr) => Math.min(min, curr))
-                : 0;
-        const count = typeof countLimit === "number"
-            ? Math.min(countLimit, inferredCount)
-            : inferredCount;
-
-        const instAttrs = attrs.filter((attr) => !!attr.divisor);
-        const instanceCount = instAttrs.length
-            ? instAttrs
-                .map((attr) => attr.count * attr.divisor)
-                .reduce((min, curr) => Math.min(min, curr))
-            : 0;
-
-        return new Attributes(
-            this.state,
-            primitive,
-            attrs,
-            count,
-            instanceCount,
-            elementBuffer,
-        );
+        return _createAttributes(this.state, elements, attributes, options);
     }
 
     /**
@@ -590,19 +427,14 @@ export class Device {
         width: number,
         height: number,
         internalFormat: F,
-        {
-            min = Filter.NEAREST,
-            mag = Filter.NEAREST,
-            wrapS = Wrap.CLAMP_TO_EDGE,
-            wrapT = Wrap.CLAMP_TO_EDGE,
-        }: TextureCreateOptions = {},
+        options?: TextureCreateOptions,
     ): Texture<F> {
-        return new Texture(
+        return _createTexture(
             this._gl,
-            width, height,
+            width,
+            height,
             internalFormat,
-            wrapS, wrapT,
-            min, mag,
+            options,
         );
     }
 
@@ -614,7 +446,8 @@ export class Device {
         image: ImageData,
         options?: TextureCreateOptions & TextureStoreOptions,
     ): Texture<InternalFormat.RGBA8> {
-        return this.createTextureWithTypedArray(
+        return _createTextureWithTypedArray(
+            this._gl,
             image.width,
             image.height,
             InternalFormat.RGBA8,
@@ -638,21 +471,18 @@ export class Device {
         data: InternalFormatToTypedArray[F],
         dataFormat: InternalFormatToDataFormat[F],
         dataType: InternalFormatToDataType[F],
-        options: TextureCreateOptions & TextureStoreOptions = {},
+        options?: TextureCreateOptions & TextureStoreOptions,
     ): Texture<F> {
-        const {
-            min = Filter.NEAREST,
-            mag = Filter.NEAREST,
-            wrapS = Wrap.CLAMP_TO_EDGE,
-            wrapT = Wrap.CLAMP_TO_EDGE,
-        } = options;
-        return new Texture(
+        return _createTextureWithTypedArray(
             this._gl,
-            width, height,
+            width,
+            height,
             internalFormat,
-            wrapS, wrapT,
-            min, mag,
-        ).store(data, dataFormat, dataType, options);
+            data,
+            dataFormat,
+            dataType,
+            options,
+        );
     }
 
     /**
@@ -660,8 +490,8 @@ export class Device {
      * depth or depth-stencil buffer with given width and height.
      *
      * Does not take ownership of provided attachments, only references them.
-     * It is still an error to use the attachments while they are written to
-     * via the framebuffer, however.
+     * WebGL will synchronize their usage so they can either be written to via
+     * the framebuffer, or written to or read via their own methods.
      */
     createFramebuffer(
         width: number,
@@ -673,177 +503,14 @@ export class Device {
             | Texture<TextureDepthInternalFormat>
             | Texture<TextureDepthStencilInternalFormat>,
     ): Framebuffer {
-        const colors = Array.isArray(color) ? color : [color];
-        assert.nonEmpty(colors, () => {
-            return "Framebuffer color attachments must not be empty";
-        });
-        colors.forEach((buffer) => {
-            assert.equal(width, buffer.width, (got, expected) => {
-                return `Expected attachment width ${expected}, got ${got}`;
-            });
-            assert.equal(height, buffer.height, (got, expected) => {
-                return `Expected attachment height ${expected}, got ${got}`;
-            });
-        });
-
-        if (depthStencil) {
-            assert.equal(width, depthStencil.width, (got, expected) => {
-                return `Expected attachment width ${expected}, got ${got}`;
-            });
-            assert.equal(height, depthStencil.height, (got, expected) => {
-                return `Expected attachment height ${expected}, got ${got}`;
-            });
-        }
-
-        return new Framebuffer(this.state, width, height, colors, depthStencil);
-    }
-}
-
-function parseDepth(
-    depth: CommandCreateOptions<void>["depth"],
-): DepthTestDescriptor | undefined {
-    if (!depth) { return undefined; }
-    assert.nonNull(depth.func, fmtParamNonNull("depth.func"));
-    return new DepthTestDescriptor(
-        depth.func || DepthFunc.LESS,
-        typeof depth.mask === "boolean" ? depth.mask : true,
-        depth.range ? depth.range[0] : 0,
-        depth.range ? depth.range[1] : 1,
-    );
-}
-
-function parseStencil(
-    stencil: CommandCreateOptions<void>["stencil"],
-): StencilTestDescriptor | undefined {
-    if (!stencil) { return undefined; }
-    assert.nonNull(stencil.func, fmtParamNonNull("stencil.func"));
-    // TODO: complete stencil validation... validation framework?
-    return new StencilTestDescriptor(
-        typeof stencil.func.func === "object"
-            ? stencil.func.func.front
-            : stencil.func.func,
-        typeof stencil.func.func === "object"
-            ? stencil.func.func.back
-            : stencil.func.func,
-        typeof stencil.func.ref !== "undefined"
-            ? typeof stencil.func.ref === "object"
-                ? stencil.func.ref.front
-                : stencil.func.ref
-            : 1,
-        typeof stencil.func.ref !== "undefined"
-            ? typeof stencil.func.ref === "object"
-                ? stencil.func.ref.back
-                : stencil.func.ref
-            : 1,
-        typeof stencil.func.mask !== "undefined"
-            ? typeof stencil.func.mask === "object"
-                ? stencil.func.mask.front
-                : stencil.func.mask
-            : 0xFF,
-        typeof stencil.func.mask !== "undefined"
-            ? typeof stencil.func.mask === "object"
-                ? stencil.func.mask.back
-                : stencil.func.mask
-            : 0xFF,
-        typeof stencil.mask !== "undefined"
-            ? typeof stencil.mask === "object"
-                ? stencil.mask.front
-                : stencil.mask
-            : 0xFF,
-        typeof stencil.mask !== "undefined"
-            ? typeof stencil.mask === "object"
-                ? stencil.mask.back
-                : stencil.mask
-            : 0xFF,
-        stencil.op
-            ? typeof stencil.op.fail === "object"
-                ? stencil.op.fail.front
-                : stencil.op.fail
-            : StencilOp.KEEP,
-        stencil.op
-            ? typeof stencil.op.fail === "object"
-                ? stencil.op.fail.back
-                : stencil.op.fail
-            : StencilOp.KEEP,
-        stencil.op
-            ? typeof stencil.op.zfail === "object"
-                ? stencil.op.zfail.front
-                : stencil.op.zfail
-            : StencilOp.KEEP,
-        stencil.op
-            ? typeof stencil.op.zfail === "object"
-                ? stencil.op.zfail.back
-                : stencil.op.zfail
-            : StencilOp.KEEP,
-        stencil.op
-            ? typeof stencil.op.zpass === "object"
-                ? stencil.op.zpass.front
-                : stencil.op.zpass
-            : StencilOp.KEEP,
-        stencil.op
-            ? typeof stencil.op.zpass === "object"
-                ? stencil.op.zpass.back
-                : stencil.op.zpass
-            : StencilOp.KEEP,
-    );
-}
-
-function parseBlend(
-    blend: CommandCreateOptions<void>["blend"],
-): BlendDescriptor | undefined {
-    if (!blend) { return undefined; }
-    assert.nonNull(blend.func, fmtParamNonNull("blend.func"));
-    assert.nonNull(blend.func.src, fmtParamNonNull("blend.func.src"));
-    assert.nonNull(blend.func.dst, fmtParamNonNull("blend.func.dst"));
-    if (typeof blend.func.src === "object") {
-        assert.nonNull(
-            blend.func.src.rgb,
-            fmtParamNonNull("blend.func.src.rgb"),
-        );
-        assert.nonNull(
-            blend.func.src.alpha,
-            fmtParamNonNull("blend.func.src.alpha"),
+        return _createFramebuffer(
+            this.state,
+            width,
+            height,
+            color,
+            depthStencil,
         );
     }
-    if (typeof blend.func.dst === "object") {
-        assert.nonNull(
-            blend.func.dst.rgb,
-            fmtParamNonNull("blend.func.dst.rgb"),
-        );
-        assert.nonNull(
-            blend.func.dst.alpha,
-            fmtParamNonNull("blend.func.dst.alpha"),
-        );
-    }
-    return new BlendDescriptor(
-        typeof blend.func.src === "object"
-            ? blend.func.src.rgb
-            : blend.func.src,
-        typeof blend.func.src === "object"
-            ? blend.func.src.alpha
-            : blend.func.src,
-        typeof blend.func.dst === "object"
-            ? blend.func.dst.rgb
-            : blend.func.dst,
-        typeof blend.func.dst === "object"
-            ? blend.func.dst.alpha
-            : blend.func.dst,
-        blend.equation
-            ? typeof blend.equation === "object"
-                ? blend.equation.rgb
-                : blend.equation
-            : BlendEquation.FUNC_ADD,
-        blend.equation
-            ? typeof blend.equation === "object"
-                ? blend.equation.alpha
-                : blend.equation
-            : BlendEquation.FUNC_ADD,
-        blend.color,
-    );
-}
-
-function fmtParamNonNull(name: string): () => string {
-    return () => `Missing parameter ${name}`;
 }
 
 function createDebugFunc(gl: any, key: string): (...args: any[]) => any {
