+++ src/element-buffer.ts
@@ -1,5 +1,6 @@
 import * as assert from "./util/assert";
-import { BufferUsage, DataType, Primitive } from "./types";
+import * as array from "./util/array";
+import { BufferUsage, DataType, Primitive, sizeOf } from "./types";
 
 export type ElementArray =
     | number[] // infers POINTS
@@ -43,6 +44,70 @@ export interface ElementBufferCreateOptions {
 
 export interface ElementBufferStoreOptions {
     offset?: number;
+}
+
+export function _createElementBuffer<T extends ElementBufferType>(
+    gl: WebGL2RenderingContext,
+    type: T,
+    primitive: Primitive,
+    size: number,
+    { usage = BufferUsage.DYNAMIC_DRAW }: ElementBufferCreateOptions = {},
+): ElementBuffer<T> {
+    return new ElementBuffer(
+        gl,
+        type,
+        primitive,
+        size,
+        size * sizeOf(type),
+        usage,
+    );
+}
+
+export function _createElementBufferWithArray(
+    gl: WebGL2RenderingContext,
+    data: ElementArray,
+    options?: ElementBufferCreateOptions,
+): ElementBuffer<DataType.UNSIGNED_INT> {
+    if (array.is2(data)) {
+        const shape = array.shape2(data);
+        assert.rangeInclusive(shape[1], 2, 3, (p) => {
+            return `Elements must be 2-tuples or 3-tuples, got ${p}-tuple`;
+        });
+        const ravel = array.ravel2(data, shape);
+        const primitive = shape[1] === 3
+            ? Primitive.TRIANGLES
+            : Primitive.LINES;
+        return _createElementBufferWithTypedArray(
+            gl,
+            DataType.UNSIGNED_INT,
+            primitive,
+            ravel,
+        );
+    }
+    return _createElementBufferWithTypedArray(
+        gl,
+        DataType.UNSIGNED_INT,
+        Primitive.POINTS,
+        data,
+        options,
+    );
+}
+
+export function _createElementBufferWithTypedArray<T extends ElementBufferType>(
+    gl: WebGL2RenderingContext,
+    type: T,
+    primitive: Primitive,
+    data: ElementBufferTypeToTypedArray[T] | number[],
+    { usage = BufferUsage.STATIC_DRAW }: ElementBufferCreateOptions = {},
+): ElementBuffer<T> {
+    return new ElementBuffer(
+        gl,
+        type,
+        primitive,
+        data.length,
+        data.length * sizeOf(type),
+        usage,
+    ).store(data);
 }
 
 /**
